requirements.txt
passlib==1.7.4
pydantic==2.6.1
python-dotenv==1.0.1
DateTime==5.4
fastapi==0.109.2
python_jose==3.3.0
SQLAlchemy==2.0.23
uvicorn==0.25.0
python-multipart==0.0.7
psycopg2-binary==2.9.9
bcrypt==4.1.3